+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.5.0] - 2021-02-25
+
 ### Added
 
 - Add `namespace` metadata validation.
@@ -154,7 +156,8 @@ since there can be a delay creating them on cluster creation.
 
 - First release
 
-[Unreleased]: https://github.com/giantswarm/app/compare/v4.4.0...HEAD
+[Unreleased]: https://github.com/giantswarm/app/compare/v4.5.0...HEAD
+[4.5.0]: https://github.com/giantswarm/app/compare/v4.4.0...v4.5.0
 [4.4.0]: https://github.com/giantswarm/app/compare/v4.3.0...v4.4.0
 [4.3.0]: https://github.com/giantswarm/app/compare/v4.2.0...v4.3.0
 [4.2.0]: https://github.com/giantswarm/app/compare/v4.1.0...v4.2.0